refactor: update parser
