+++ cmd/root.go
@@ -26,7 +26,7 @@ import (
 	"log"
 	"os"
 
-	"github.com/snamiki1212/go-gen-slice-accessors/internal"
+	"github.com/snamiki1212/go-gen-slice-accessors/internal/parser"
 	"github.com/snamiki1212/go-gen-slice-accessors/internal/pluralizer"
 	"github.com/snamiki1212/go-gen-slice-accessors/internal/reader"
 	"github.com/snamiki1212/go-gen-slice-accessors/internal/writer"
@@ -39,14 +39,14 @@ var rootCmd = &cobra.Command{
 	Short: "Generate accessors for each field in the slice struct.",
 	RunE: func(cmd *cobra.Command, _ []string) error {
 		// Load arguments
-		if err := internal.Args.Load(); err != nil {
+		if err := parser.Args.Load(); err != nil {
 			return fmt.Errorf("loader error: %w", err)
 		}
 
 		// Parse source code
-		generator, err := internal.
-			NewParser(reader.NewReader(internal.Args.Input), pluralizer.NewPluralizer()).
-			Parse(internal.Args)
+		generator, err := parser.
+			NewParser(reader.NewReader(parser.Args.Input), pluralizer.NewPluralizer()).
+			Parse(parser.Args)
 		if err != nil {
 			return fmt.Errorf("parse error: %w", err)
 		}
@@ -58,7 +58,7 @@ var rootCmd = &cobra.Command{
 		}
 
 		// Write to output file
-		err = writer.Write(internal.Args.Output, txt)
+		err = writer.Write(parser.Args.Output, txt)
 		if err != nil {
 			return fmt.Errorf("write error: %w", err)
 		}
@@ -77,27 +77,27 @@ func Execute() {
 
 func init() {
 	// entity
-	rootCmd.Flags().StringVarP(&internal.Args.Entity, "entity", "e", "", "[required] Target entity name")
+	rootCmd.Flags().StringVarP(&parser.Args.Entity, "entity", "e", "", "[required] Target entity name")
 	_ = rootCmd.MarkFlagRequired("entity")
 
 	// slice
-	rootCmd.Flags().StringVarP(&internal.Args.Slice, "slice", "s", "", "[required] Target slice name")
+	rootCmd.Flags().StringVarP(&parser.Args.Slice, "slice", "s", "", "[required] Target slice name")
 	_ = rootCmd.MarkFlagRequired("slice")
 
 	// input
-	rootCmd.Flags().StringVarP(&internal.Args.Input, "input", "i", "", "[required] Input file name")
+	rootCmd.Flags().StringVarP(&parser.Args.Input, "input", "i", "", "[required] Input file name")
 	_ = rootCmd.MarkFlagRequired("input")
 
 	// output
-	rootCmd.Flags().StringVarP(&internal.Args.Output, "output", "o", "", "[required] Output file name")
+	rootCmd.Flags().StringVarP(&parser.Args.Output, "output", "o", "", "[required] Output file name")
 	_ = rootCmd.MarkFlagRequired("output")
 
 	// fieldNamesToExclude
-	rootCmd.Flags().StringSliceVarP(&internal.Args.FieldNamesToExclude, "exclude", "x", []string{}, "Field names to exclude")
+	rootCmd.Flags().StringSliceVarP(&parser.Args.FieldNamesToExclude, "exclude", "x", []string{}, "Field names to exclude")
 
 	// rename
-	rootCmd.Flags().StringSliceVarP(&internal.Args.RawRenames, "rename", "r", []string{}, "Rename accessor name \n e.g. --rename=Name:GetName")
+	rootCmd.Flags().StringSliceVarP(&parser.Args.RawRenames, "rename", "r", []string{}, "Rename accessor name \n e.g. --rename=Name:GetName")
 
 	// import
-	rootCmd.Flags().StringSliceVarP(&internal.Args.RawImportPaths, "import", "m", []string{}, "Import path name \n e.g. --import=time \n e.g. --import=time:aliasTime")
+	rootCmd.Flags().StringSliceVarP(&parser.Args.RawImportPaths, "import", "m", []string{}, "Import path name \n e.g. --import=time \n e.g. --import=time:aliasTime")
 }
