+++ internal/parser/arguments.go
@@ -1,4 +1,4 @@
-package internal
+package parser
 
 import (
 	"fmt"
