+++ internal/parser/parser_test.go
@@ -1,4 +1,4 @@
-package internal
+package parser
 
 import (
 	"go/ast"
